LPS-59594 Refactor package names
